chore: split api_9 test
