+++ packages/amplify-e2e-tests/src/__tests__/api_9b.test.ts
@@ -38,75 +38,6 @@ describe('amplify add api (GraphQL)', () => {
     deleteProjectDir(projRoot);
   });
 
-  it('init a project and add the simple_model api with multiple authorization providers', async () => {
-    const appName = createRandomName();
-    await initJSProjectWithProfile(projRoot, { name: appName });
-    await addApiWithoutSchema(projRoot, { transformerVersion: 1 });
-    await updateApiSchema(projRoot, appName, 'simple_model.graphql');
-    await updateApiWithMultiAuth(projRoot, {});
-    await amplifyPush(projRoot);
-
-    const meta = getProjectMeta(projRoot);
-    const { output } = meta.api[appName];
-    const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
-    const { graphqlApi } = await getAppSyncApi(GraphQLAPIIdOutput, meta.providers.awscloudformation.Region);
-
-    expect(graphqlApi).toBeDefined();
-    expect(graphqlApi.authenticationType).toEqual('API_KEY');
-    expect(graphqlApi.additionalAuthenticationProviders).toHaveLength(3);
-    expect(graphqlApi.additionalAuthenticationProviders).toHaveLength(3);
-
-    const cognito = graphqlApi.additionalAuthenticationProviders.filter(a => a.authenticationType === 'AMAZON_COGNITO_USER_POOLS')[0];
-
-    expect(cognito).toBeDefined();
-    expect(cognito.userPoolConfig).toBeDefined();
-
-    const iam = graphqlApi.additionalAuthenticationProviders.filter(a => a.authenticationType === 'AWS_IAM')[0];
-
-    expect(iam).toBeDefined();
-
-    const oidc = graphqlApi.additionalAuthenticationProviders.filter(a => a.authenticationType === 'OPENID_CONNECT')[0];
-
-    expect(oidc).toBeDefined();
-    expect(oidc.openIDConnectConfig).toBeDefined();
-    expect(oidc.openIDConnectConfig.issuer).toEqual('https://facebook.com/');
-    expect(oidc.openIDConnectConfig.clientId).toEqual('clientId');
-    expect(oidc.openIDConnectConfig.iatTTL).toEqual(1000);
-    expect(oidc.openIDConnectConfig.authTTL).toEqual(2000);
-
-    expect(GraphQLAPIIdOutput).toBeDefined();
-    expect(GraphQLAPIEndpointOutput).toBeDefined();
-    expect(GraphQLAPIKeyOutput).toBeDefined();
-
-    expect(graphqlApi).toBeDefined();
-    expect(graphqlApi.apiId).toEqual(GraphQLAPIIdOutput);
-  });
-
-  it('init a project and add the simple_model api, match transformer version to current version', async () => {
-    const name = `simplemodel${generateRandomShortId()}`;
-    await initJSProjectWithProfile(projRoot, { name });
-    await addApiWithoutSchema(projRoot, { transformerVersion: 1 });
-    await updateApiSchema(projRoot, name, 'simple_model.graphql');
-    await amplifyPush(projRoot);
-
-    const meta = getProjectMeta(projRoot);
-    const { output } = meta.api[name];
-    const { GraphQLAPIIdOutput, GraphQLAPIEndpointOutput, GraphQLAPIKeyOutput } = output;
-    const { graphqlApi } = await getAppSyncApi(GraphQLAPIIdOutput, meta.providers.awscloudformation.Region);
-
-    expect(GraphQLAPIIdOutput).toBeDefined();
-    expect(GraphQLAPIEndpointOutput).toBeDefined();
-    expect(GraphQLAPIKeyOutput).toBeDefined();
-
-    expect(graphqlApi).toBeDefined();
-    expect(graphqlApi.apiId).toEqual(GraphQLAPIIdOutput);
-
-    const transformConfig = getTransformConfig(projRoot, name);
-    expect(transformConfig).toBeDefined();
-    expect(transformConfig.Version).toBeDefined();
-    expect(transformConfig.Version).toEqual(TRANSFORM_CURRENT_VERSION);
-  });
-
   it('init a project with a simple model , add a function and removes the dependent @model', async () => {
     const projectName = `simplemodel${generateRandomShortId()}`;
     const nextSchema = 'initial_key_blog.graphql';
